backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai>=0.8.3
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.12
